Rename & move NodeEdge and NodeEdgeApplyResult out of public interface
